refactored create account screen
refactored create account screen
